refactor(db): stage checkpoint attempt lifecycle
