+++ crates/laminar-db/src/checkpoint_coordinator/mod.rs
@@ -4,6 +4,8 @@
 
 mod allocation;
 mod artifacts;
+mod attempt;
+mod attempt_failure;
 mod capture;
 mod commit_index;
 #[cfg(feature = "cluster")]
@@ -23,7 +25,9 @@ mod subscription_output;
 use std::collections::{BTreeMap, HashMap, HashSet};
 use std::num::NonZeroU32;
 use std::sync::Arc;
-use std::time::{Duration, Instant};
+use std::time::Duration;
+#[cfg(feature = "cluster")]
+use std::time::Instant;
 
 use bytes::Bytes;
 use futures::{stream::FuturesUnordered, StreamExt};
@@ -49,7 +53,9 @@ use tracing::warn;
 
 use crate::error::DbError;
 pub(crate) use allocation::EpochAllocator;
-use allocation::{checked_successor_epoch, require_canonical_attempt, SinkEpochReservation};
+#[cfg(feature = "cluster")]
+use allocation::SinkEpochReservation;
+use allocation::{checked_successor_epoch, require_canonical_attempt};
 use retention::{run_gc_worker, GcRequest};
 
 const MAX_RETENTION_IO_CONCURRENCY: usize = 8;
@@ -668,798 +674,6 @@ impl CheckpointCoordinator {
     #[must_use]
     pub fn store(&self) -> &dyn CheckpointStore {
         self.store.as_ref()
-    }
-
-    async fn record_outcome_until(
-        &self,
-        attempt: CheckpointAttempt,
-        verdict: laminar_core::checkpoint_decision::CheckpointVerdict,
-        committed_checkpoint: Option<CommittedCheckpointRef>,
-        assignment_fence: Option<laminar_core::checkpoint::CheckpointAssignmentFence>,
-        leader_proof: Option<LeaderProof>,
-        deadline: tokio::time::Instant,
-    ) -> Result<laminar_core::checkpoint_decision::CheckpointOutcome, DbError> {
-        use laminar_core::checkpoint_decision::RecordOutcomeResult;
-
-        #[cfg(feature = "cluster")]
-        let result = if let Some(controller) = self.cluster_controller.as_ref() {
-            let proof = leader_proof
-                .as_ref()
-                .ok_or_else(|| DbError::Checkpoint("cluster outcome has no leader proof".into()))?;
-            let fence = assignment_fence.clone().ok_or_else(|| {
-                DbError::Checkpoint("cluster outcome has no assignment fence".into())
-            })?;
-            let authority = controller.checkpoint_authority().map_err(|error| {
-                DbError::Checkpoint(format!("cluster checkpoint authority: {error}"))
-            })?;
-            tokio::time::timeout_at(
-                deadline,
-                authority.record_cluster_outcome(
-                    proof,
-                    attempt.epoch,
-                    attempt.checkpoint_id,
-                    fence,
-                    verdict.clone(),
-                    committed_checkpoint.clone(),
-                ),
-            )
-            .await
-            .map_err(|_| DbError::Checkpoint("cluster outcome create timed out".into()))?
-            .map_err(|error| DbError::Checkpoint(format!("cluster outcome create: {error}")))?
-        } else {
-            let store = self.decision_store.as_ref().ok_or_else(|| {
-                DbError::Checkpoint("checkpoint outcome requires a decision store".into())
-            })?;
-            tokio::time::timeout_at(
-                deadline,
-                store.record_outcome(
-                    attempt.epoch,
-                    attempt.checkpoint_id,
-                    CheckpointScope::Local,
-                    None,
-                    None,
-                    verdict.clone(),
-                    committed_checkpoint.clone(),
-                ),
-            )
-            .await
-            .map_err(|_| DbError::Checkpoint("checkpoint outcome create timed out".into()))?
-            .map_err(|error| DbError::Checkpoint(format!("checkpoint outcome create: {error}")))?
-        };
-        #[cfg(not(feature = "cluster"))]
-        let result = {
-            let store = self.decision_store.as_ref().ok_or_else(|| {
-                DbError::Checkpoint("checkpoint outcome requires a decision store".into())
-            })?;
-            tokio::time::timeout_at(
-                deadline,
-                store.record_outcome(
-                    attempt.epoch,
-                    attempt.checkpoint_id,
-                    CheckpointScope::Local,
-                    None,
-                    None,
-                    verdict.clone(),
-                    committed_checkpoint.clone(),
-                ),
-            )
-            .await
-            .map_err(|_| DbError::Checkpoint("checkpoint outcome create timed out".into()))?
-            .map_err(|error| DbError::Checkpoint(format!("checkpoint outcome create: {error}")))?
-        };
-
-        let winner = match result {
-            RecordOutcomeResult::Created(outcome) | RecordOutcomeResult::Unchanged(outcome) => {
-                outcome
-            }
-            RecordOutcomeResult::Conflict { winner } => winner,
-        };
-        if winner.epoch != attempt.epoch
-            || winner.checkpoint_id != attempt.checkpoint_id
-            || winner.verdict != verdict
-            || winner.committed_checkpoint != committed_checkpoint
-            || winner.assignment_fence != assignment_fence
-            || winner.leader_proof != leader_proof
-            || winner.deployment_id != self.expected_deployment_id()?
-        {
-            return Err(DbError::Checkpoint(format!(
-                "checkpoint {} lost its immutable outcome race",
-                attempt.checkpoint_id
-            )));
-        }
-        Ok(winner)
-    }
-
-    async fn abort_attempt_until(
-        &mut self,
-        attempt: CheckpointAttempt,
-        assignment_fence: Option<laminar_core::checkpoint::CheckpointAssignmentFence>,
-        leader_proof: Option<LeaderProof>,
-        deadline: tokio::time::Instant,
-    ) -> Result<(), DbError> {
-        self.phase = CheckpointPhase::Deciding;
-        #[cfg(feature = "cluster")]
-        let cluster_scope = self.cluster_controller.is_some();
-        #[cfg(not(feature = "cluster"))]
-        let cluster_scope = false;
-        self.record_outcome_until(
-            attempt,
-            laminar_core::checkpoint_decision::CheckpointVerdict::Abort,
-            None,
-            assignment_fence,
-            leader_proof,
-            deadline,
-        )
-        .await?;
-        let cleanup_deadline = tokio::time::Instant::now() + self.config.cleanup_timeout;
-        let rollback = self
-            .rollback_sinks_until(attempt.epoch, cleanup_deadline)
-            .await;
-        let witness_cleanup = if rollback.is_ok() {
-            self.clear_sink_witness_until(cleanup_deadline).await
-        } else {
-            Ok(())
-        };
-        let artifact_cleanup = if cluster_scope {
-            self.failure_requires_recovery = true;
-            Ok(())
-        } else {
-            self.cleanup_local_checkpoint_artifacts_until(attempt, cleanup_deadline)
-                .await
-        };
-        rollback?;
-        witness_cleanup?;
-        artifact_cleanup?;
-        self.allocator.advance_epoch_to(checked_successor_epoch(
-            attempt.epoch,
-            "closing an aborted checkpoint",
-        )?);
-        self.phase = CheckpointPhase::Idle;
-        Ok(())
-    }
-
-    fn failed_result(
-        &mut self,
-        attempt: CheckpointAttempt,
-        started: Instant,
-        error: String,
-        disposition: CheckpointFailureDisposition,
-    ) -> CheckpointResult {
-        let duration = started.elapsed();
-        self.phase = CheckpointPhase::Idle;
-        self.record_checkpoint_outcome(false, attempt, duration, None);
-        if disposition == CheckpointFailureDisposition::RequiresRecovery {
-            self.failure_requires_recovery = true;
-        }
-        CheckpointResult {
-            success: false,
-            checkpoint_id: attempt.checkpoint_id,
-            epoch: attempt.epoch,
-            duration,
-            error: Some(error),
-            failure_disposition: Some(disposition),
-        }
-    }
-
-    async fn fail_before_commit(
-        &mut self,
-        attempt: CheckpointAttempt,
-        started: Instant,
-        error: DbError,
-        flags: u64,
-        assignment_fence: Option<laminar_core::checkpoint::CheckpointAssignmentFence>,
-        leader_proof: Option<LeaderProof>,
-        _attempt_deadline: tokio::time::Instant,
-        sink_epoch_publication: SinkEpochPublication,
-    ) -> CheckpointResult {
-        #[cfg(not(feature = "cluster"))]
-        let _ = flags;
-        // Once an exact attempt has been reserved, failure settlement owns a private cleanup
-        // budget. The attempt deadline fences new capture/durable work, but must not cancel the
-        // durable Abort and rollback that make the attempt terminal.
-        let cleanup_deadline = tokio::time::Instant::now() + self.config.cleanup_timeout;
-        let mut message = error.to_string();
-        if let Err(seal) = self
-            .seal_sink_epoch_until(attempt.epoch, cleanup_deadline)
-            .await
-        {
-            self.failure_requires_recovery = true;
-            message = format!("{message}; {seal}");
-        }
-        match self
-            .abort_attempt_until(
-                attempt,
-                assignment_fence.clone(),
-                leader_proof.clone(),
-                cleanup_deadline,
-            )
-            .await
-        {
-            Ok(()) => {
-                // Durable Abort is already terminal. Bound its best-effort cluster hint under a
-                // fresh, explicit cleanup window rather than reviving the expired attempt deadline
-                // or allowing notification I/O to hang.
-                #[cfg(feature = "cluster")]
-                if let Some(controller) = self.cluster_controller.as_ref() {
-                    let notification_deadline =
-                        tokio::time::Instant::now() + self.config.cleanup_timeout;
-                    publish_terminal_hint_until(
-                        notification_deadline,
-                        controller.announce_barrier(&BarrierAnnouncement {
-                            epoch: attempt.epoch,
-                            checkpoint_id: attempt.checkpoint_id,
-                            assignment_fence,
-                            leader_proof,
-                            phase: Phase::Abort,
-                            flags,
-                        }),
-                    )
-                    .await;
-                }
-                // A slow best-effort hint must not consume the local successor epoch's required
-                // continuation budget.
-                let continuation_deadline =
-                    tokio::time::Instant::now() + self.config.cleanup_timeout;
-                let requires_recovery = self.failure_requires_recovery;
-                let successor = if !requires_recovery && self.has_checkpoint_committable_sinks() {
-                    self.begin_sink_epoch_until(continuation_deadline, sink_epoch_publication)
-                        .await
-                        .err()
-                } else {
-                    None
-                };
-                let (error, disposition) = match (requires_recovery, successor) {
-                    (true, _) => (message, CheckpointFailureDisposition::RequiresRecovery),
-                    (false, Some(successor)) => (
-                        format!("{message}; successor sink epoch failed: {successor}"),
-                        CheckpointFailureDisposition::RequiresRecovery,
-                    ),
-                    (false, None) => (message, CheckpointFailureDisposition::Retryable),
-                };
-                self.failed_result(attempt, started, error, disposition)
-            }
-            Err(abort) => self.failed_result(
-                attempt,
-                started,
-                format!("{message}; durable Abort or rollback failed: {abort}"),
-                CheckpointFailureDisposition::RequiresRecovery,
-            ),
-        }
-    }
-
-    async fn run_checkpoint_attempt(
-        &mut self,
-        request: CheckpointRequest,
-        attempt: CheckpointAttempt,
-        quorum: QuorumStage,
-        started: Instant,
-        deadline: tokio::time::Instant,
-        sink_epoch_publication: SinkEpochPublication,
-    ) -> Result<CheckpointResult, DbError> {
-        require_canonical_attempt(attempt, "checkpoint admission")?;
-        let (flags, assignment_fence, terminal_handoff) = (
-            request.flags,
-            request.assignment_fence.clone(),
-            sink_epoch_admission::is_terminal_handoff(
-                request.flags,
-                request.handoff_replay_pending,
-            ),
-        );
-        #[cfg(feature = "cluster")]
-        let validation_proof = match &quorum {
-            QuorumStage::Captured { leader_proof, .. } => Some(leader_proof.clone()),
-            QuorumStage::RunInline => self
-                .cluster_controller
-                .as_ref()
-                .and_then(|controller| controller.capture_leader_proof()),
-        };
-        #[cfg(not(feature = "cluster"))]
-        let validation_proof = None;
-        if self.failure_requires_recovery {
-            return Ok(self
-                .fail_before_commit(
-                    attempt,
-                    started,
-                    DbError::Checkpoint(
-                        "a prior checkpoint has unresolved durable or sink state".into(),
-                    ),
-                    flags,
-                    assignment_fence,
-                    validation_proof,
-                    deadline,
-                    sink_epoch_publication,
-                )
-                .await);
-        }
-        if tokio::time::Instant::now() >= deadline {
-            return Ok(self
-                .fail_before_commit(
-                    attempt,
-                    started,
-                    DbError::Checkpoint("checkpoint deadline expired before durable work".into()),
-                    flags,
-                    assignment_fence,
-                    validation_proof,
-                    deadline,
-                    sink_epoch_publication,
-                )
-                .await);
-        }
-        if let Err(error) = self.validate_request(&request) {
-            return Ok(self
-                .fail_before_commit(
-                    attempt,
-                    started,
-                    error,
-                    flags,
-                    assignment_fence,
-                    validation_proof,
-                    deadline,
-                    sink_epoch_publication,
-                )
-                .await);
-        }
-
-        let assignment_fence = request.assignment_fence.clone();
-        #[cfg(feature = "cluster")]
-        let (scope, leader_proof, quorum_watermark) =
-            if let Some(controller) = self.cluster_controller.clone() {
-                let fence = assignment_fence.as_ref().ok_or_else(|| {
-                    DbError::Checkpoint("cluster checkpoint has no assignment fence".into())
-                })?;
-                let (proof, participants, cluster_watermark) = match quorum {
-                    QuorumStage::RunInline => {
-                        return Ok(self
-                            .fail_before_commit(
-                                attempt,
-                                started,
-                                DbError::Checkpoint(
-                                    "cluster checkpoint reached durable execution without a \
-                                 precomputed certified quorum"
-                                        .into(),
-                                ),
-                                flags,
-                                assignment_fence,
-                                validation_proof,
-                                deadline,
-                                sink_epoch_publication,
-                            )
-                            .await);
-                    }
-                    QuorumStage::Captured {
-                        cluster_watermark,
-                        participants,
-                        leader_proof,
-                    } => (leader_proof, participants, cluster_watermark),
-                };
-                if let Err(error) =
-                    self.validate_captured_quorum(&controller, fence, participants, &proof)
-                {
-                    return Ok(self
-                        .fail_before_commit(
-                            attempt,
-                            started,
-                            error,
-                            flags,
-                            assignment_fence,
-                            Some(proof),
-                            deadline,
-                            sink_epoch_publication,
-                        )
-                        .await);
-                }
-                (
-                    CheckpointScope::Cluster,
-                    Some(proof),
-                    Some(cluster_watermark),
-                )
-            } else {
-                let _ = quorum;
-                (CheckpointScope::Local, None, None)
-            };
-        #[cfg(not(feature = "cluster"))]
-        let (scope, leader_proof, quorum_watermark) = {
-            let _ = quorum;
-            (CheckpointScope::Local, None, None)
-        };
-
-        self.phase = CheckpointPhase::PreCommitting;
-        let descriptors = match self.pre_commit_sinks_until(attempt.epoch, deadline).await {
-            Ok(descriptors) => descriptors,
-            Err(error) => {
-                return Ok(self
-                    .fail_before_commit(
-                        attempt,
-                        started,
-                        error,
-                        flags,
-                        assignment_fence,
-                        leader_proof,
-                        deadline,
-                        sink_epoch_publication,
-                    )
-                    .await);
-            }
-        };
-        let packed = match self
-            .pack_checkpoint(attempt, request, descriptors, deadline)
-            .await
-        {
-            Ok(packed) => packed,
-            Err(error) => {
-                return Ok(self
-                    .fail_before_commit(
-                        attempt,
-                        started,
-                        error,
-                        flags,
-                        assignment_fence,
-                        leader_proof,
-                        deadline,
-                        sink_epoch_publication,
-                    )
-                    .await);
-            }
-        };
-        let local_manifest_bytes = match self.persist_checkpoint_until(&packed, deadline).await {
-            Ok(bytes) => bytes,
-            Err(error) => {
-                return Ok(self
-                    .fail_before_commit(
-                        attempt,
-                        started,
-                        error,
-                        flags,
-                        assignment_fence,
-                        leader_proof,
-                        deadline,
-                        sink_epoch_publication,
-                    )
-                    .await);
-            }
-        };
-
-        let manifests = match self
-            .await_prepared_participant_manifests(
-                attempt,
-                assignment_fence.as_ref(),
-                (packed.manifest.clone(), local_manifest_bytes),
-                deadline,
-            )
-            .await
-        {
-            Ok(manifests) => manifests,
-            Err(error) => {
-                return Ok(self
-                    .fail_before_commit(
-                        attempt,
-                        started,
-                        error,
-                        flags,
-                        assignment_fence,
-                        leader_proof,
-                        deadline,
-                        sink_epoch_publication,
-                    )
-                    .await);
-            }
-        };
-        let predecessor = match self
-            .authoritative_committed_predecessor_until(scope, deadline)
-            .await
-        {
-            Ok(predecessor) => predecessor,
-            Err(error) => {
-                return Ok(self
-                    .fail_before_commit(
-                        attempt,
-                        started,
-                        error,
-                        flags,
-                        assignment_fence,
-                        leader_proof,
-                        deadline,
-                        sink_epoch_publication,
-                    )
-                    .await);
-            }
-        };
-        let predecessor_source_watermarks = match self
-            .predecessor_source_watermarks_until(predecessor.as_ref(), deadline)
-            .await
-        {
-            Ok(source_watermarks) => source_watermarks,
-            Err(error) => {
-                return Ok(self
-                    .fail_before_commit(
-                        attempt,
-                        started,
-                        error,
-                        flags,
-                        assignment_fence,
-                        leader_proof,
-                        deadline,
-                        sink_epoch_publication,
-                    )
-                    .await);
-            }
-        };
-        let index = match self
-            .build_validated_committed_index_until(
-                attempt,
-                scope,
-                assignment_fence.clone(),
-                predecessor.clone(),
-                &predecessor_source_watermarks,
-                &manifests,
-                quorum_watermark,
-                deadline,
-            )
-            .await
-        {
-            Ok(index) => index,
-            Err(error) => {
-                return Ok(self
-                    .fail_before_commit(
-                        attempt,
-                        started,
-                        error,
-                        flags,
-                        assignment_fence,
-                        leader_proof,
-                        deadline,
-                        sink_epoch_publication,
-                    )
-                    .await);
-            }
-        };
-        #[cfg(feature = "cluster")]
-        let subscription_commit_stats = (scope == CheckpointScope::Cluster)
-            .then(|| subscription_output::subscription_commit_stats(&manifests));
-        #[cfg(all(debug_assertions, feature = "cluster"))]
-        checkpoint_kill_gate(
-            "leader",
-            attempt,
-            predecessor
-                .as_ref()
-                .map(|reference| (reference.checkpoint_id, reference.epoch)),
-        )
-        .await;
-        let reference = match self.create_committed_index_until(&index, deadline).await {
-            Ok(reference) => reference,
-            Err(error) => {
-                return Ok(self
-                    .fail_before_commit(
-                        attempt,
-                        started,
-                        error,
-                        flags,
-                        assignment_fence,
-                        leader_proof,
-                        deadline,
-                        sink_epoch_publication,
-                    )
-                    .await);
-            }
-        };
-
-        self.phase = CheckpointPhase::Deciding;
-        #[cfg(feature = "cluster")]
-        let commit_visibility_started = Instant::now();
-        let outcome = self
-            .record_outcome_until(
-                attempt,
-                laminar_core::checkpoint_decision::CheckpointVerdict::Commit,
-                Some(reference.clone()),
-                assignment_fence.clone(),
-                leader_proof.clone(),
-                deadline,
-            )
-            .await;
-        if let Err(error) = outcome {
-            return Ok(self.failed_result(
-                attempt,
-                started,
-                format!("commit outcome is in-doubt: {error}"),
-                CheckpointFailureDisposition::RequiresRecovery,
-            ));
-        }
-        #[cfg(feature = "cluster")]
-        self.record_subscription_commit(
-            subscription_commit_stats,
-            commit_visibility_started.elapsed(),
-            attempt,
-        );
-
-        let predecessor_checkpoint_id = index
-            .predecessor
-            .as_ref()
-            .map_or(0, |reference| reference.checkpoint_id);
-        self.last_committed_ref = Some(reference.clone());
-        self.last_committed_source_watermarks = Some((reference, index.source_watermarks.clone()));
-        self.last_committed_manifest = manifests
-            .iter()
-            .find(|(manifest, _)| manifest.participant_id == self.store.participant_id())
-            .map(|(manifest, _)| Arc::new(manifest.clone()));
-        self.prepared.remove(&attempt);
-        self.allocator.advance_epoch_to(checked_successor_epoch(
-            attempt.epoch,
-            "closing a committed checkpoint",
-        )?);
-        #[cfg(feature = "cluster")]
-        if let Some(controller) = self.cluster_controller.as_ref() {
-            controller
-                .publish_committed_checkpoint_progress(
-                    &index.channel_progress,
-                    &index.source_watermarks,
-                )
-                .map_err(DbError::Checkpoint)?;
-            // The durable Commit is already immutable. Its cluster hint is best-effort and must
-            // not delay sink continuation or the terminal caller reply without bound.
-            let notification_deadline = tokio::time::Instant::now() + self.config.cleanup_timeout;
-            publish_terminal_hint_until(
-                notification_deadline,
-                controller.announce_barrier(&BarrierAnnouncement {
-                    epoch: attempt.epoch,
-                    checkpoint_id: attempt.checkpoint_id,
-                    assignment_fence: assignment_fence.clone(),
-                    leader_proof: leader_proof.clone(),
-                    phase: Phase::Commit,
-                    flags,
-                }),
-            )
-            .await;
-        }
-
-        let continuation_deadline = tokio::time::Instant::now() + self.config.cleanup_timeout;
-        let fencing_token = leader_proof.as_ref().map_or(1, |proof| proof.fencing_token);
-        let participant_manifests = manifests
-            .iter()
-            .map(|(manifest, _)| manifest)
-            .collect::<Vec<_>>();
-        let continuation = self
-            .commit_external_sinks_until(
-                attempt,
-                &participant_manifests,
-                fencing_token,
-                predecessor_checkpoint_id,
-                continuation_deadline,
-            )
-            .await;
-        let continuation = self
-            .continue_committed_sink_epoch_until(
-                continuation,
-                &index,
-                leader_proof.as_ref(),
-                terminal_handoff,
-                continuation_deadline,
-                sink_epoch_publication,
-            )
-            .await;
-
-        let duration = started.elapsed();
-        self.phase = CheckpointPhase::Idle;
-        let checkpoint_bytes = self
-            .last_committed_manifest
-            .as_ref()
-            .map(|manifest| manifest.node_data.object_length);
-        self.record_checkpoint_outcome(true, attempt, duration, checkpoint_bytes);
-        let continuation_error = continuation.err().map(|error| {
-            self.failure_requires_recovery = true;
-            format!(
-                "checkpoint {} committed, but sink continuation requires recovery: {error}",
-                attempt.checkpoint_id
-            )
-        });
-        Ok(CheckpointResult {
-            success: true,
-            checkpoint_id: attempt.checkpoint_id,
-            epoch: attempt.epoch,
-            duration,
-            error: continuation_error,
-            failure_disposition: None,
-        })
-    }
-
-    pub async fn checkpoint(
-        &mut self,
-        request: CheckpointRequest,
-    ) -> Result<CheckpointResult, DbError> {
-        let started = Instant::now();
-        let deadline = tokio::time::Instant::from_std(started) + self.config.checkpoint_timeout;
-        #[cfg(feature = "cluster")]
-        let local = self.cluster_controller.is_none();
-        #[cfg(not(feature = "cluster"))]
-        let local = true;
-        if !local {
-            return Err(DbError::Checkpoint(
-                "cluster checkpoints require reserved pipeline admission and certified Prepare"
-                    .into(),
-            ));
-        }
-        let attempt = self.allocate_attempt_until(deadline).await?;
-        if let Err(error) = self
-            .begin_checkpoint_artifacts_until(attempt, None, None, deadline)
-            .await
-        {
-            return Ok(self
-                .fail_before_commit(
-                    attempt,
-                    started,
-                    error,
-                    request.flags,
-                    request.assignment_fence.clone(),
-                    None,
-                    deadline,
-                    SinkEpochPublication::Immediate,
-                )
-                .await);
-        }
-        self.run_checkpoint_attempt(
-            request,
-            attempt,
-            QuorumStage::RunInline,
-            started,
-            deadline,
-            SinkEpochPublication::Immediate,
-        )
-        .await
-    }
-
-    pub async fn checkpoint_with_offsets(
-        &mut self,
-        request: CheckpointRequest,
-    ) -> Result<CheckpointResult, DbError> {
-        self.checkpoint(request).await
-    }
-
-    pub(crate) async fn checkpoint_preallocated_started(
-        &mut self,
-        request: CheckpointRequest,
-        attempt: CheckpointAttempt,
-        quorum: QuorumStage,
-        started: Instant,
-        deadline: tokio::time::Instant,
-    ) -> Result<CheckpointResult, DbError> {
-        self.run_checkpoint_attempt(
-            request,
-            attempt,
-            quorum,
-            started,
-            deadline,
-            SinkEpochPublication::DeferredToTail,
-        )
-        .await
-    }
-
-    pub(crate) async fn abandon_epoch_until(
-        &mut self,
-        checkpoint_id: u64,
-        epoch: u64,
-        error: String,
-        flags: u64,
-        assignment_fence: Option<laminar_core::checkpoint::CheckpointAssignmentFence>,
-        leader_proof: Option<LeaderProof>,
-        deadline: tokio::time::Instant,
-        sink_epoch_publication: SinkEpochPublication,
-    ) -> Result<CheckpointResult, DbError> {
-        let attempt = require_canonical_attempt(
-            CheckpointAttempt::new(epoch, checkpoint_id),
-            "checkpoint abandonment",
-        )?;
-        let started = Instant::now();
-        Ok(self
-            .fail_before_commit(
-                attempt,
-                started,
-                DbError::Checkpoint(error),
-                flags,
-                assignment_fence,
-                leader_proof,
-                deadline,
-                sink_epoch_publication,
-            )
-            .await)
     }
 
     #[cfg(feature = "cluster")]
