+++ tools/readability-check/function-baseline.tsv
@@ -42,7 +42,6 @@ crates/laminar-db/src/aggregate_state/checkpoints/ipc.rs	validate_ipc_batch_layo
 crates/laminar-db/src/aggregate_state/checkpoints/ipc.rs	preflight_scalar_ipc_restore	194	hot-path-local-exception	The remaining body owns one cohesive record/batch/state iteration or accounting boundary; extraction without a dedicated benchmark could add calls, allocation, copying, or obscure exact state mutation order.	Add or run a focused benchmark, then extract only cold validation, preparation, error conversion, or result assembly; this function must not grow.
 crates/laminar-db/src/aggregate_state/compile.rs	push_aggregate	132	hot-path-local-exception	The remaining body owns one cohesive record/batch/state iteration or accounting boundary; extraction without a dedicated benchmark could add calls, allocation, copying, or obscure exact state mutation order.	Add or run a focused benchmark, then extract only cold validation, preparation, error conversion, or result assembly; this function must not grow.
 crates/laminar-db/src/builder/mod.rs	build	187	characterization-required-local-exception	The function is known readability debt, but its observable branches are not isolated enough for a safe semantic split in this pass; wrapper-only extraction would not improve comprehension.	Characterize observable outcomes and extract a domain-owned phase with explicit inputs and outputs; this function must not grow.
-crates/laminar-db/src/checkpoint_coordinator/mod.rs	run_checkpoint_attempt	424	module-review-baseline	Checkpoint admission, persistence, cleanup, and completion remain one protocol implementation.	Extract attempt lifecycle and cleanup owners under existing tests; file may not grow.
 crates/laminar-db/src/coordinated_recovery/mod.rs	run	202	module-review-baseline	Recovery round phases, quorum evidence, and lifecycle fencing share exact ordering.	Split immutable phase preparation from protocol execution; file may not grow.
 crates/laminar-db/src/coordinated_recovery/mod.rs	observe	275	module-review-baseline	Recovery round phases, quorum evidence, and lifecycle fencing share exact ordering.	Split immutable phase preparation from protocol execution; file may not grow.
 crates/laminar-db/src/coordinated_recovery/mod.rs	observe_nonparticipant_release	154	module-review-baseline	Recovery round phases, quorum evidence, and lifecycle fencing share exact ordering.	Split immutable phase preparation from protocol execution; file may not grow.
